YouTube Shuffle
YouTube Shuffle
This is a javascript application for listening to music on YouTube. The idea is that you amass a collection of songs you like, and YouTube Shuffle plays them at random. YouTube playlists are complicated and the user interface is complicated and bad. When you hear a song you like, just paste its URL into YouTube Shuffle.

There is no server required to host this page -- it can be downloaded and run locally. The playlist is passively saved in HTML5 local storage, but can also be batch imported/exported for easy sharing (e.g. using pastebin) with other people or across devices. A copy of the page is hosted on GitHub: http://jbergknoff.github.io/youtube-shuffle.